fire hover events during repaint of HexPanel
I'm not sure whether this should stay that way. It simplifies some things
and makes the GUI more responsive but generates more events and some UI
artifacts:

The reason for this feature is to make the marker follow the mouse during
scrolling. However, the marker sometimes visibly lags behind during fast
scrolls but is corrected immediately.